working on the quiz rendering
