+++ config/openaiclient.js
@@ -1,6 +1,7 @@
+require("dotenv").config({ path: "./config/.env" });
 const Openai = require('openai');
 
-const openai = new OpenAI({
+const openai = new Openai({
   apiKey: process.env.OPENAI_API_KEY,
 });
