+++ package.json
@@ -5,7 +5,8 @@
   "main": "server.js",
   "scripts": {
     "test": "jest",
-    "start": "node server.js"
+    "start": "node server.js",
+    "fast" : "nodemon server.js"
   },
   "repository": {
     "type": "git",
