+++ views/index.ejs
@@ -8,19 +8,27 @@
 <body>
 <header class="">
     <nav class="">
-        <div>
-            <img src="logo.png" alt="ThriveTribe Logo" class="">
+        <div class="">
+            <a href="/">
+                <img src="logo.png" alt="ThriveTribe Logo" class="">
+            </a>
         </div>
         <div class="">
-            <a href="#" class="">about</a>
-            <a href="#" class="">quiz</a>
-            <a href="#" class="">types</a>
-            <a href="#" class="">find</a>
-            <a href="#" class="">1-0-1</a>
-            <a href="#" class="">blog</a>
-            <a href="#" class="">therapist sign in</a>
+            <a href="#" class="">About</a>
+            <a href="/quiz" class="">Quiz</a>
+            <a href="#" class="">Types</a>
+            <a href="#" class="">Find A Therapist</a>
+            <a href="#" class="">Therapy 101</a>
+            <a href="#" class="">Blog</a>
+            <a href="#" class="">Therapist Sign In</a>
         </div>
     </nav>
+    <section class="class">
+        <div class="class">
+            <h1>Welcome to ThriveTribe, Your Path to Culturally Attuned Therapy</h1>
+            <p>ThriveTribe is your compass in the quest for mental wellness, specially designed for Black women, including immigrants and children of immigrants. Here, you'll find a community that values your unique journey, connecting you with therapists who not only understand your cultural background but also speak your language, literally and metaphorically. Begin with our personalized quiz, tailored to guide you to therapy options that resonate with your story. At ThriveTribe, your cultural identity is not just acknowledged—it's celebrated as a cornerstone of your healing journey.</p>
+        </div>
+    </section>
 </header>
     
 </body>
