+++ views/quiz.ejs
@@ -7,29 +7,39 @@
 </head>
 <body>
     <main>
-        <h1>What's your therapy type?</h1>
-        <div id="quiz-container">
-            <% if (sessionData.currentStep !== "end") { %>
-                <!-- Display the question for the current step -->
-                <p><%= sessionData.question %></p>
-                
-                <!-- Display the answers for the current step -->
-                <% if (sessionData.answers && sessionData.answers.length > 0) { %>
-                    <% sessionData.answers.forEach(function(answer) { %>
-                        <form action="/quiz/handleQuizStep" method="POST">
-                            <button type="submit" name="userInput" value="<%= answer.text %>"><%= answer.text %></button>
-                        </form>
-                    <% }); %>
-                <% } else { %>
-                    <p>Click the button to start the quiz.</p>
+        <div class="container">
+            <section>
+                <h1>Welcome to Your Personal Therapy Exploration Quiz!</h1>
+                <p>Embarking on a journey toward mental wellness is a brave and transformative step. This quiz is designed to be your companion as you explore the therapeutic paths that might resonate with your unique experiences and emotional landscape.</p>
+                <p>Our goal is to introduce you to various therapy styles through a series of thoughtful questions. By reflecting on your preferences, coping mechanisms, and emotional tendencies, we aim to provide you with insights that could guide you toward the most compatible therapeutic approaches.</p>
+                <p>Ready to Begin? Dive in with an open heart and an open mind. Your journey to deeper self-understanding starts here. When you're ready, hit the "Take Quiz" button, and let's explore the therapeutic possibilities together.</p>
+                <div class="button">
                     <form action="/quiz/handleQuizStep" method="POST">
-                        <button type="submit" name="userInput" value="Start Quiz">Start Quiz</button>
-                    </form>
-                <% } %>
-            <% } else { %>
-                <!-- Display the final message -->
-                <p><%= sessionData.message %></p>
-            <% } %>
+                        <button type="submit" name="action" value="takeQuiz">Take Quiz</button>
+                    </form>  
+                </div>
+            </section>
+        </div>
+
+    
+        <div id="quiz-container">
+            <% if (stepData && stepData.interMessage) { %>
+                <h2><%= stepData.interMessage %></h2>
+               <% if (stepData.question) { %>
+                 <h2><%= stepData.question %></h2>
+                 <ul>
+                   <% stepData.answers.forEach(answer => { %>
+                     <li>
+                       <form action="/quiz/handleQuizStep" method="POST">
+                         <button type="submit" name="userInput" value="<%= answer.text %>">
+                           <%= answer.text %>
+                         </button>
+                       </form>
+                     </li>
+                   <% }); %>
+                 </ul>
+               <% } %>
+            <% } %>  
         </div>
     </main>
 </body>
